pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "MarketScope"
version = "0.1.0"
description = "行业分析AI应用"
authors = [{name = "Developer", email = "developer@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.0.352",
    "langchain-community>=0.0.38",
    "langchain-openai>=0.0.8",
    "openai>=1.10.0",
    "tavily-python>=0.3.3",
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "pandas>=2.1.4",
    "numpy>=1.26.2",
    "pydantic>=2.5.2",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.23.2",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["MarketScope*"]